- Moved logic files closer to the actual components. - Add support for signal prop on the song image component.
