+++ src/renderer/src/components/song/song-detail/SongDetail.tsx
@@ -1,11 +1,10 @@
 import { Component } from "solid-js";
 import Bar from "../../bar/Bar";
-import { seek } from "../../../lib/Music";
+import { seek, duration, song, timestamp } from "@renderer/components/song/song.utils";
 import formatTime from "../../../lib/time-formatter";
 import SongImage from "../SongImage";
 import SongControls from "./SongControls";
 import { isSongUndefined } from "../../../lib/song";
-import { duration, song, timestamp } from "@renderer/lib/state/song";
 import "./styles.css";
 
 const SongDetail: Component = () => {
