+++ src/renderer/src/components/song/song-item/SongItem.tsx
@@ -2,7 +2,7 @@ import { Component, createSignal, onMount } from "solid-js";
 import { ResourceID, Song } from "../../../../../@types";
 import { ignoreClickInContextMenu } from "../context-menu/SongContextMenu";
 import draggable from "../../../lib/draggable/draggable";
-import { song as selectedSong } from "../../../lib/state/song";
+import { song as selectedSong } from "../../../components/song/song.utils";
 import SongHint from "../SongHint";
 import SongImage from "../SongImage";
 import "./styles.css";
